+++ app/src/main/java/com/hanbitkang/mypokemon/MainActivity.kt
@@ -3,13 +3,7 @@ package com.hanbitkang.mypokemon
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.*
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import com.hanbitkang.mypokemon.ui.MpApp
-import com.hanbitkang.core_designsystem.theme.MyPokemonTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
