+++ core-designsystem/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.hanbitkang.core_designsystem">
+    package="com.hanbitkang.core.designsystem">
 
 </manifest>
